Fix bugs in ProcessAlarms, and knock-on bug in SchedulerTest that relied on this particular misbehavior.  Basically, we should block the requested time if there are no outstanding alarms, and be woken up when an alarm arrives.  Also fixed mistreatment of quiescence in mock_scheduler.cc that was exposed by the other bug fixes.  Finally, re-factor blocking class from rewrite_driver into scheduler as we need to use it in various other places (including in tests).  (jmaessen)
Declare independence from the tyranny of Apache pool destruction semantics by 
passing NULL for the pool into the Serf constructor.

+++ src/net/instaweb/rewriter/image_combine_filter.cc
@@ -147,7 +147,7 @@ class SpriteFuture {
   // Set x_px and y_px to the alignment for this image/div combination
   // before spriting.
   bool SetAlignmentValues(Css::Value* x_value, Css::Value* y_value,
-                          size_t image_width, size_t image_height,
+                          int image_width, int image_height,
                           int* x_px, int* y_px) {
     bool ret = true;
     if (x_value->GetLexicalUnitType() == Css::Value::NUMBER) {
